Override server provided defaults with parameters passed in the URL if they match the data type
