+++ client/js/socket-events/configuration.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const $ = require("jquery");
+const URI = require("urijs");
 const socket = require("../socket");
 const templates = require("../../views");
 const options = require("../options");
@@ -84,4 +85,38 @@ socket.on("configuration", function(data) {
 				$(this).data("lastvalue", nick);
 			});
 	});
+
+	if ($(document.body).hasClass("public")) {
+		const params = URI(document.location.search).search(true);
+
+		for (let key in params) {
+			// Support `channels` as a compatibility alias with other clients
+			if (key === "channels") {
+				key = "join";
+			}
+
+			if (!data.defaults.hasOwnProperty(key)) {
+				continue;
+			}
+
+			let value = params[key];
+
+			if (key === "join") {
+				value = value.split(",").map((chan) => {
+					if (!chan.match(/^[#&!+]/)) {
+						return `#${chan}`;
+					}
+
+					return chan;
+				}).join(", ");
+			}
+
+			// Override server provided defaults with parameters passed in the URL if they match the data type
+			switch (typeof data.defaults[key]) {
+			case "boolean": data.defaults[key] = value === "1" || value === "true"; break;
+			case "number": data.defaults[key] = Number(value); break;
+			case "string": data.defaults[key] = String(value); break;
+			}
+		}
+	}
 });
